refactor: ManagerCafeService 명령 조회 분리
